+++ tiptap/index.js
@@ -43,6 +43,7 @@ export { CarveKit } from './carve-kit.js';
 // Individual extensions
 export { CarveInsert } from './extensions/carve-insert.js';
 export { CarveDelete } from './extensions/carve-delete.js';
+export { CarveCriticComment } from './extensions/carve-critic-comment.js';
 export { CarveDiv } from './extensions/carve-div.js';
 export { CarveMath } from './extensions/carve-math.js';
 export { CarveFootnoteDefinition } from './extensions/carve-footnote-definition.js';